Convert LogNormal maxValue from Long to Double to prevent loss of precision causing errors.

+++ src/main/java/com/github/tomakehurst/wiremock/client/ResponseDefinitionBuilder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2011-2024 Thomas Akehurst
+ * Copyright (C) 2011-2025 Thomas Akehurst
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -157,7 +157,7 @@ public class ResponseDefinitionBuilder {
   }
 
   public ResponseDefinitionBuilder withLogNormalRandomDelay(
-      double medianMilliseconds, double sigma, Long maxValue) {
+      double medianMilliseconds, double sigma, Double maxValue) {
     return withRandomDelay(new LogNormal(medianMilliseconds, sigma, maxValue));
   }
